frontend Fixes #7
