+++ package.json
@@ -10,6 +10,7 @@
     "vercel-dev": "vercel env pull -y && vercel dev --local-config vercel.dev.json"
   },
   "dependencies": {
+    "animate.css": "^4.1.0",
     "axios": "^0.19.2",
     "copy-to-clipboard": "^3.3.1",
     "core-js": "^3.6.5",
