+++ src/components/Footer.vue
@@ -1,0 +1,69 @@
+<template>
+  <footer class="text-center">
+    <span>
+      Font in use
+      <a
+        class="hover:underline text-gray-600"
+        href="https://www.fontspring.com/fonts/los-andes-type/fibra/fibra-light"
+        title="fibra-light"
+      >
+        fibra-light
+      </a>
+      designed by
+      <a
+        class="hover:underline text-gray-600"
+        href="https://losandestype.myportfolio.com/"
+        title="Los Andes Type"
+      >
+        Los Andes Type</a
+      >
+      <br class="sm:hidden" />
+      and licensed under
+      <a
+        class="hover:underline text-gray-600"
+        href="https://www.fontspring.com//lic/jcefupvyrh"
+        title="Fontspring"
+      >
+        Custom EULA
+      </a>
+    </span>
+    <span>
+      Icons made by
+      <a
+        class="hover:underline text-gray-600"
+        href="https://www.flaticon.com/authors/freepik"
+        title="Freepik"
+        >Freepik</a
+      >
+      and
+      <a
+        class="hover:underline  text-gray-600"
+        href="https://www.flaticon.com/authors/pixel-perfect"
+        title="Pixel perfect"
+        >Pixel perfect</a
+      >
+      from
+      <a
+        class="hover:underline  text-gray-600"
+        href="https://www.flaticon.com/"
+        title="Flaticon"
+        >www.flaticon.com</a
+      >
+    </span>
+    <span>
+      Font Awesome by Dave Gandy -
+      <a
+        class="hover:underline  text-gray-600"
+        href="http://fontawesome.io"
+        title="Font Awesome"
+        >http://fontawesome.io</a
+      >
+    </span>
+  </footer>
+</template>
+
+<script>
+export default {};
+</script>
+
+<style></style>
