+++ src/components/Logo.vue
@@ -1,0 +1,31 @@
+<template>
+  <header class="flex flex-row items-center justify-center">
+    <svg
+      class="h-full fill-current text-indigo-500"
+      xmlns="http://www.w3.org/2000/svg"
+      viewBox="0 0 480.026 480.026"
+    >
+      <path
+        d="M475.922 229.325l-144-160c-3.072-3.392-7.36-5.312-11.904-5.312h-96c-6.304 0-12.032 3.712-14.624 9.472-2.56 5.792-1.504 12.544 2.72 17.216l134.368 149.312-134.368 149.28c-4.224 4.704-5.312 11.456-2.72 17.216 2.592 5.792 8.32 9.504 14.624 9.504h96c4.544 0 8.832-1.952 11.904-5.28l144-160c5.472-6.08 5.472-15.36 0-21.408z"
+      />
+      <path
+        d="M267.922 229.325l-144-160c-3.072-3.392-7.36-5.312-11.904-5.312h-96c-6.304 0-12.032 3.712-14.624 9.472-2.56 5.792-1.504 12.544 2.72 17.216l134.368 149.312L4.114 389.293c-4.224 4.704-5.312 11.456-2.72 17.216 2.592 5.792 8.32 9.504 14.624 9.504h96c4.544 0 8.832-1.952 11.904-5.28l144-160c5.472-6.08 5.472-15.36 0-21.408z"
+      />
+    </svg>
+    <svg
+      class="h-full ml-6 fill-current text-gray-700"
+      xmlns="http://www.w3.org/2000/svg"
+      viewBox="0 0 100 30"
+    >
+      <path
+        d="M28.303837 16.331874h-11.93355v2.503912h9.269808c-.639293 4.581633-4.288615 7.565019-9.829194 7.565019-6.020049 0-11.267617-4.501716-11.267617-11.400801 0-6.8990852 5.247568-11.4008098 11.267617-11.4008098 5.407392 0 8.9768 3.356314 9.616099 8.0178608h2.74365c-.559385-5.9401392-5.540578-10.6016869-12.359749-10.6016869-7.618293 0-14.224367 5.6204931-14.224367 13.9047209 0 8.390779 6.606074 14.064543 14.224367 14.064543 7.485106 0 12.492936-4.821363 12.492936-11.667176zm3.862425 2.503912c0 5.647129 4.208707 10.148846 10.148846 10.148846 5.886861 0 10.095569-4.501717 10.095569-10.148846 0-5.540577-4.208708-9.9890222-10.095569-9.9890222-5.940139 0-10.148846 4.4484452-10.148846 9.9890222zm2.770285.07992c0-4.261986 3.063304-7.644935 7.378561-7.644935 4.288623 0 7.298647 3.382949 7.298647 7.644935 0 4.208706-3.010024 7.644927-7.298647 7.644927-4.315257 0-7.378561-3.436221-7.378561-7.644927zM75.798318 1.3882935H54.781407v2.6903768h8.950167V28.611707h3.01003V4.0786703h9.056714zm2.344088 17.4474925c0 5.647129 4.235345 10.148846 10.148848 10.148846 5.886862 0 10.122212-4.501717 10.122212-10.148846 0-5.540577-4.23535-9.9890222-10.122212-9.9890222-5.913503 0-10.148848 4.4484452-10.148848 9.9890222zm2.770286.07992c0-4.261986 3.063303-7.644935 7.378562-7.644935 4.288622 0 7.325282 3.382949 7.325282 7.644935 0 4.208706-3.03666 7.644927-7.325282 7.644927-4.315259 0-7.378562-3.436221-7.378562-7.644927z"
+      />
+    </svg>
+  </header>
+</template>
+
+<script>
+export default {};
+</script>
+
+<style></style>
