+++ src/components/Scissors.vue
@@ -1,0 +1,52 @@
+<template>
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    viewBox="-22 0 512 512.00033"
+    v-on:click="$emit('click')"
+  >
+    <path
+      d="M195.14499 279.64862l-16.72657-28.45312c-2.52734-4.30078-6.30468-7.73047-10.82812-9.83594L31.97311 178.1994l15.82422-33.10937 433.457031 206.7539c1.066407.51172 1.109375 2.01172.074219 2.58204l-29.527344 16.20312c-5.804687 3.18359-12.71875 3.64844-18.898437 1.26953L207.96921 290.30878c-5.38672-1.95313-9.92188-5.72656-12.82422-10.66016zm0 0"
+      fill="#c7cfe2"
+    />
+    <path
+      d="M195.20358 233.20331l-16.79687 28.41016c-2.53907 4.29296-6.32422 7.71484-10.85157 9.80859L31.78561 334.24628l15.74219 33.14844 433.960936-205.69141c1.070313-.50781 1.113282-2.00781.082032-2.57813l-29.488282-16.27734c-5.796875-3.19922-12.710937-3.67969-18.894531-1.31641L208.05124 222.5705c-5.39063 1.9414-9.9336 5.70312-12.84766 10.63281zm0 0"
+      fill="#e4eaf7"
+    />
+    <path
+      d="M100.2153 474.5705c40.10937-19.00782 57.19531-66.89844 38.16406-106.96485-19.02734-40.0625-66.96875-57.13281-107.07422-38.125C-8.80033 348.49237-25.88626 396.383-6.85501 436.4455c19.02734 40.0664 66.96484 57.13672 107.07031 38.125zm-69.30078-56.02735c-9.13281-19.22265-.9336-42.20703 18.3125-51.32812 19.24609-9.125 42.25-.92969 51.38281 18.29297 9.13281 19.22656.93359 42.20703-18.3125 51.33203-19.24609 9.12109-42.25.92969-51.38281-18.29688zm0 0"
+      fill="#01c8f0"
+    />
+    <path
+      d="M100.2153 474.5705c31.76562-15.05469 49.07031-48.22657 45.44531-81.27735-2.95312 26.76172-19.30469 51.42188-45.44531 63.8086-40.10547 19.01172-88.04297 1.9414-107.07031-38.125-3.96094-8.33203-6.33203-17.00391-7.28516-25.6836-1.57812 14.30078.65625 29.19922 7.28516 43.15235 19.02734 40.0664 66.96484 57.13672 107.07031 38.125zm0 0"
+      fill="#01a3dc"
+    />
+    <path
+      d="M100.60983 385.508c1.20313 2.53906 2.09766 5.14062 2.71484 7.76562 1.90625-8.1875 1.17188-17.04687-2.71484-25.23047-9.13281-19.22656-32.13672-27.41797-51.38281-18.29687-16.70703 7.92187-25.08203 26.28125-21.02735 43.5664 2.57422-11.04296 9.97266-20.85937 21.02735-26.09765 19.24609-9.1211 42.25-.92969 51.38281 18.29297zm0 0"
+      fill="#01a3dc"
+    />
+    <path
+      d="M171.64499 159.56659c18.41406-38.39453 2.17968-84.429686-36.25391-102.824218L97.11764 38.425966C58.68405 20.031435 12.59811 36.250185-5.81204 74.644716c-18.41407 38.394532-2.17969 84.429684 36.2539 102.824214l38.27344 18.31641c38.43359 18.39453 84.51953 2.17969 102.92969-36.21875zm-88.60938 1.66406L49.78952 145.3205c-19.21485-9.19531-27.32813-32.207034-18.125-51.39844 9.20703-19.195312 32.23828-27.300781 51.45312-18.105469l33.2461 15.910157c19.21093 9.195312 27.32421 32.207032 18.12109 51.398442-9.20313 19.19531-32.23828 27.30078-51.44922 18.10546zm0 0"
+      fill="#01c8f0"
+    />
+    <path
+      d="M171.64499 159.56659c6.61328-13.78906 8.73437-28.5625 6.98828-42.691405-1.00391 8.058595-3.28125 16.097655-6.98828 23.828125-18.41016 38.39453-64.49219 54.60937-102.92969 36.21484l-38.27344-18.3164c-24.6289-11.78516-40.11719-34.92578-43.23828-60.128909-3.96875 31.980469 12.54297 64.304689 43.23828 78.996089l38.27344 18.31641c38.43359 18.39453 84.51953 2.17969 102.92969-36.21875zm0 0"
+      fill="#01a3dc"
+    />
+    <path
+      d="M116.36374 91.726748c10.7539 5.144531 18.02343 14.621093 20.73437 25.332031 4.42969-17.429688-3.8789-36.132813-20.73437-44.199219l-33.2461-15.910156c-19.21484-9.195313-42.25-1.089844-51.45312 18.105468-4.05078 8.449219-4.74219 17.640626-2.61328 26.066407.62109-2.4375 1.48437-4.847657 2.61328-7.203125 9.20703-19.191406 32.23828-27.296875 51.45312-18.101563zm0 0"
+      fill="#01a3dc"
+    />
+    <path
+      d="M103.43405 481.35956c21.21094-10.05469 37.23047-27.75781 45.11328-49.85156 7.88281-22.09375 6.67969-45.92969-3.38281-67.1211-4.77735-10.05468-11.23047-18.78125-18.8711-26.01562l83.3125-39.48828 220.671879 80.05468c8.246093 3.14454 17.402343 2.51954 25.140625-1.72656l29.511718-16.19531c2.96875-1.625 4.761719-4.73047 4.660157-8.11328-.09375-3.375-2.046875-6.37891-5.101563-7.83594L339.84811 276.0744c-3.74609-1.78515-8.22656-.19531-10.01172 3.54688-1.78515 3.74219-.19922 8.22656 3.54297 10.01172l133.968751 63.89843-19.152343 10.51172c-3.875 2.125-8.46875 2.4336-12.601563.84766-.04297-.0156-.08594-.0312-.132813-.0508L228.69967 289.8322l52.71485-24.98438 20.70312 9.87109c3.74219 1.78516 8.22266.19922 10.00781-3.54687 1.78516-3.73828.19922-8.22266-3.54296-10.01172l-9.66797-4.60937 185.800779-88.06639c3.054687-1.45313 5.0156248-4.45703 5.10937475-7.83594.08984375-3.40625-1.68359375-6.48437-4.62890575-8.10156l-29.480469-16.26563c-7.71875-4.27343-16.875-4.92578-25.136719-1.79687l-220.875 79.51172-50.375-24.02344c7.09766-6.66016 13.11328-14.69922 17.56641-23.98437l2.99218-6.23829c9.37891-19.55859 10.57422-41.593745 3.36719-62.050776-7.20703-20.457032-21.94922-36.878906-41.51172-46.242188l-44.59765-21.34375c-40.33594-19.300781-88.875-2.203125-108.21094 38.113282l-2.98828 6.234374c-5.33203 11.125-8.01953 23.050782-8.01953 35.03125 0 9.082032 1.54297 18.199218 4.65234 27.019528 7.20313 20.45703 21.94531 36.87891 41.51172 46.24219l44.60156 21.34375c2.88281 1.37891 5.8086 2.55469 8.76172 3.56641l86.97266 40.5039c3.14843 1.46485 5.7539 3.83203 7.51953 6.82813l.82422 1.40234-.81641 1.38281c-1.78125 3.00391-4.39062 5.35938-7.55078 6.82032L56.01999 314.7619c-9.44532 1.04297-18.875 3.64063-27.92969 7.9336-21.21094 10.05468-37.23047 27.75781-45.11328 49.84765-3.44141 9.64063-5.14844 19.60547-5.14844 29.54297 0 12.84375 2.85547 25.63672 8.53125 37.58203 20.77344 43.74219 73.29297 62.44141 117.07422 41.69141zM30.57467 169.20722c-15.94531-7.62891-27.96093-21.01172-33.83203-37.6836-5.87109-16.671872-4.89453-34.628904 2.7461-50.566404l2.99218-6.238281c15.75391-32.855469 55.3125-46.789063 88.17969-31.058594l44.59766 21.34375c15.94531 7.628907 27.96094 21.011719 33.83203 37.683594 5.87109 16.671875 4.89453 34.628905-2.74609 50.566405l-2.99219 6.23438c-15.75391 32.85937-55.3125 46.78906-88.17969 31.0625zm140.19141 65.34375l-51.59375-24.02735c9.42188-1.83984 18.46484-5.34765 26.69141-10.33203l47.83593 22.8125c-1.90234 1.89844-3.57812 4.03516-4.96093 6.375l-7.61328 12.88281c-2.88282-3.23047-6.38282-5.85547-10.35938-7.71093zm14.10938 30.88281l16.78906-28.41016c2.02344-3.42187 5.1914-6.03906 8.92969-7.38281l225.132808-81.04297c.04297-.0195.08984-.0352.132812-.0508 4.136719-1.58203 8.722656-1.26172 12.589844.87891l19.132812 10.55861L278.2778 249.71503c-.0195.008-.043.0195-.0625.0273l-72.17188 34.20703c-.004.004-.008.004-.008.004l-93.0039 44.08203c-7.98438-5.07422-16.73438-8.83203-25.875-11.13281l83.55097-38.66802c5.92578-2.73828 10.82031-7.16016 14.16797-12.80078zM-.07376 433.22675c-8.34375-17.56641-9.33594-37.32032-2.80469-55.63282 6.53516-18.3164 19.81641-32.99218 37.40234-41.32421 36.30078-17.21094 79.85157-1.70704 97.07422 34.55859 8.33985 17.5625 9.33594 37.32031 2.80469 55.63281-6.53516 18.3125-19.81641 32.98828-37.40234 41.32422-36.30469 17.20703-79.84766 1.70313-97.07422-34.55859zm0 0"
+    />
+    <path
+      d="M109.16842 417.48847c4.13282-11.58594 3.50391-24.08985-1.77343-35.19922-10.89453-22.9336-38.42969-32.73828-61.38282-21.85938-6.60156 3.12891-12.42578 7.82813-16.83593 13.59375-2.52344 3.29297-1.89844 8.00781 1.39453 10.52735 3.29297 2.52343 8.00781 1.89843 10.52734-1.39454 3.01172-3.93359 6.82813-7.01562 11.34766-9.15625 15.47656-7.33593 34.03906-.72656 41.38281 14.73047 3.55469 7.48438 3.97656 15.90625 1.19531 23.71094-2.78515 7.80469-8.44922 14.05859-15.9414 17.61328-15.47657 7.33594-34.04297.72656-41.38282-14.73047-1.57812-3.32031-2.54687-6.83593-2.8789-10.45312-.37891-4.12891-4.03516-7.16797-8.16797-6.78906-3.89453.35937-6.82031 3.63281-6.82031 7.46875 0 .23046.0117.46484.0312.69921.49609 5.37891 1.92968 10.59766 4.26562 15.51563 10.89453 22.93359 38.42969 32.73828 61.38281 21.85937 11.1211-5.26953 19.52344-14.55078 23.65625-26.13671zm0 0M140.49655 147.95722l1.49609-3.11719c5.1211-10.67969 5.77344-22.714845 1.83985-33.88672-3.9375-11.175781-11.98828-20.144531-22.67188-25.257812L84.75046 68.273622C62.72311 57.730654 36.21139 67.070498 25.6528 89.090029l-1.49219 3.113281c-2.91406 6.074219-4.38281 12.589844-4.38281 19.132812 0 4.960938.84375 9.9375 2.54297 14.757818 3.93359 11.17187 11.98437 20.14062 22.67187 25.2539l9.83594 4.70703c3.74219 1.79297 8.22656.21094 10.01563-3.52734 1.79296-3.74219.21093-8.22656-3.53125-10.01563l-9.83985-4.71093c-7.0625-3.38282-12.38672-9.3125-14.98828-16.699222-2.60156-7.386719-2.16797-15.34375 1.21875-22.402344l1.49219-3.117188c6.98047-14.558594 24.50781-20.730468 39.07422-13.761718l36.40625 17.425781c7.0664 3.378906 12.38672 9.308593 14.98828 16.695312 2.60547 7.382809 2.17187 15.343749-1.21485 22.402349l-1.49609 3.11718c-6.98047 14.5586-24.50781 20.73438-39.07031 13.76172-3.74219-1.78906-8.22656-.20703-10.01563 3.53516-1.78906 3.73437-.20703 8.22265 3.53125 10.01562 22.03125 10.54297 48.53907 1.20313 59.09766-20.8164zm0 0M225.78952 255.62518c-.0899-4.33984-3.6836-7.78906-8.03125-7.69921-4.34375.0898-7.79297 3.68359-7.70313 8.02343.0899 4.33985 3.6836 7.78516 8.02735 7.69532 4.34765-.0898 7.79687-3.67969 7.70703-8.01954zm0 0"
+    />
+  </svg>
+</template>
+
+<script>
+export default {};
+</script>
+
+<style></style>
